+++ apps/prairielearn/src/components/FileBrowser.tsx
@@ -603,7 +603,7 @@ function DirectoryBrowserBody({
                         output: f.sync_warnings,
                       })
                     : ''}
-                <span><i class="far fa-file-alt fa-fw"></i></span>
+                <span><i class="far fa-file-alt"></i></span>
                 ${f.canView
                   ? html`<a href="${paths.urlPrefix}/file_view/${encodePath(f.path)}">${f.name}</a>`
                   : html`<span>${f.name}</span>`}
@@ -689,7 +689,7 @@ function DirectoryBrowserBody({
           (d) => html`
             <tr>
               <td colspan="2">
-                <i class="fa fa-folder fa-fw"></i>
+                <i class="fa fa-folder"></i>
                 ${d.canView
                   ? html`<a href="${paths.urlPrefix}/file_view/${encodePath(d.path)}">${d.name}</a>`
                   : html`<span>${d.name}</span>`}