+++ packages/compiled-assets/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prairielearn/compiled-assets
 
+## 3.2.4
+
+### Patch Changes
+
+- b55261c: Upgrade to TypeScript 5.9
+- Updated dependencies [b55261c]
+  - @prairielearn/html@4.0.18
+
+## 3.2.3
+
+### Patch Changes
+
+- 0edbdcb: Upgrade more JavaScript dependencies
+- 23adb05: Upgrade all JavaScript dependencies
+- Updated dependencies [23adb05]
+  - @prairielearn/html@4.0.17
+
 ## 3.2.2
 
 ### Patch Changes